Reorganized recurrences serializer
